// client/src/app/_models/city-json.ts
export class CityJSON {
    'place_id': 234854017;
    'licence': 'Data © OpenStreetMap contributors, ODbL 1.0. https://osm.org/copyright';
    'osm_type': 'relation';
    'osm_id': 206542;
    'boundingbox': ['37.1838918', '37.2769328', '-80.4806781', '-80.3791226'];
    'lat': '37.2296566';
    'lon': '-80.4136767';
    'display_name': 'Blacksburg, Montgomery County, Virginia, United States of America';
    'class': 'boundary';
    'type': 'administrative';
    'importance': 0.572716953241034;
    'icon': 'https://nominatim.openstreetmap.org/images/mapicons/poi_boundary_administrative.p.20.png';
    'address': {
      'town': 'Blacksburg',
      'county': 'Montgomery County',
      'state': 'Virginia',
      'country': 'United States of America',
      'country_code': 'us'
    };
}